requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
pandas>=2.0.0
faiss-cpu>=1.7.4
torch>=2.0.0
transformers>=4.30.0
Pillow>=10.0.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
sentencepiece>=0.1.99
protobuf>=4.24.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
aiofiles>=23.1.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
